Make footer smaller (#25)
It works for the homepage.

There is an issue in signup page, the root div doesn't expand to the
bottom of the page.

I don't know if this is the case in other pages.

I know it is not an issue with the footer since it is at the very bottom
of the root div.

ie. the root div not reaching bottom of page is a different issue.